1.6-NVIDIA-VIM_PROJECT/main.py
from openai import OpenAI
import os 
from dotenv import load_dotenv

load_dotenv()

nvidia_api_key = os.getenv("NVIDIA_VIM_API_KEY")

client = OpenAI(
  base_url = "https://integrate.api.nvidia.com/v1",
  api_key = nvidia_api_key
)

completion = client.chat.completions.create(
  model="meta/llama3-70b-instruct",
  messages=[{"role":"user","content":"Provide me an essay on CNN(Convolutional Neural Networks)"}],
  temperature=0.5,
  top_p=1,
  max_tokens=1024,
  stream=True
)

for chunk in completion:
  if not getattr(chunk, "choices", None):
    continue
  if chunk.choices[0].delta.content is not None:
    print(chunk.choices[0].delta.content, end="")
