1.6-NVIDIA-VIM_PROJECT/pyproject.toml
[project]
name = "1-6-nvidia-vim-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-nvidia-ai-endpoints>=1.0.4",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.21.0",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
